+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
 
   # ui - widget
   after_layout: ^1.2.0
+  flutter_riverpod: ^2.5.1
 
   # ui - util
   flutter_quill: ^9.2.1
